+++ dsms/knowledge/groups/__init__.py
@@ -2,8 +2,8 @@
 
 from .models import BaseGroup, Group, GroupList, GroupListBase, User, UserList
 from .public import (
-    EXTERNALLY_PUBLIC_GROUP,
-    INTERNALLY_PUBLIC_GROUP,
+    INTERNAL_GROUP,
+    PUBLIC_GROUP,
     refresh_public_groups,
 )
 
@@ -11,8 +11,8 @@ __all__ = [
     "Group",
     "GroupList",
     "GroupListBase",
-    "INTERNALLY_PUBLIC_GROUP",
-    "EXTERNALLY_PUBLIC_GROUP",
+    "INTERNAL_GROUP",
+    "PUBLIC_GROUP",
     "refresh_public_groups",
     "User",
     "BaseGroup",